feat: carry the assurance campaign, and delete the rewrite it left half-built
The campaign's own work is here: the compiler-surfaces package compiling under
--all-targets, the clippy policy call sites, the gate and policy plumbing, the
fixture-wired seam attribution root fix, and the load-sensitive test races.

What is removed is the one thing that did not land.

snapshot::Options gained a rewrites field, ManifestRewrite, and a rewrite()
that redirects a manifest declaration by replacing `path = "…"` in the copied
text. Nothing ever constructed a ManifestRewrite, and workspace.rs never named
the field, so the crate did not compile. The two lossy-text findings and the
four comment findings in cargo xtask lints were all in that code.

Text is the wrong layer for it. A declaration that climbs out of the tree fails
to resolve in the copy because the copy places the tree and what it reads
beside it by two independent prefix substitutions that agree only when the
allowed directory is a sibling of the root. Rewriting the manifest patches the
symptom and cannot see single quotes, inline tables, [patch], an inherited
[workspace.dependencies] entry, a .cargo/config.toml paths entry, or a build.rs
reading ../. The placement is what has to become one substitution.

Outside.declared goes with it: it was added to feed the rewrite, it is read
nowhere, and its doc claimed the path as the manifest writes it while cargo
metadata reports the resolved absolute one.

+++ crates/rust-mutants-cli/src/report/candidates.rs
@@ -13,6 +13,7 @@ pub const SCHEMA_VERSION: u32 = 1;
 
 /// Every place the selected rules target, as the walk found them.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
 pub struct CandidatesDocument {
     /// [`DOCUMENT_TYPE`].
     pub document_type: String,
@@ -21,13 +22,14 @@ pub struct CandidatesDocument {
     /// The engine that walked the tree.
     pub tool_version: String,
     /// How many candidates the document holds, which is the number `catalog` will rule on.
-    pub count: u32,
+    pub count: usize,
     /// Every candidate, in the order the walk found them.
     pub candidates: Vec<CandidateDocument>,
 }
 
 /// One place a rule targets.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
 pub struct CandidateDocument {
     /// The identity the catalog would give it, when the walk could compute one.
     pub id: String,
